+++ src/mcp_coder/workflows/implement/core.py
@@ -5,7 +5,6 @@ prerequisites checking, task tracker preparation, and task processing loops.
 """
 
 import logging
-import sys
 from pathlib import Path
 from typing import Optional
 
@@ -29,58 +28,6 @@ PR_INFO_DIR = "pr_info"
 
 # Setup logger
 logger = logging.getLogger(__name__)
-
-
-def resolve_project_dir(project_dir_arg: Optional[str]) -> Path:
-    """Convert project directory argument to absolute Path, with validation.
-
-    Args:
-        project_dir_arg: Optional project directory path string, uses current directory if None
-
-    Returns:
-        Path: Validated absolute path to project directory
-
-    Raises:
-        SystemExit: If path is invalid, doesn't exist, not a directory, not accessible, or not a git repo
-    """
-    # Use current directory if no argument provided
-    if project_dir_arg is None:
-        project_path = Path.cwd()
-    else:
-        project_path = Path(project_dir_arg)
-
-    # Resolve to absolute path
-    try:
-        project_path = project_path.resolve()
-    except (OSError, ValueError) as e:
-        logger.error(f"Invalid project directory path: {e}")
-        sys.exit(1)
-
-    # Validate directory exists
-    if not project_path.exists():
-        logger.error(f"Project directory does not exist: {project_path}")
-        sys.exit(1)
-
-    # Validate it's a directory
-    if not project_path.is_dir():
-        logger.error(f"Project path is not a directory: {project_path}")
-        sys.exit(1)
-
-    # Validate directory is accessible
-    try:
-        # Test read access by listing directory
-        list(project_path.iterdir())
-    except PermissionError:
-        logger.error(f"No read access to project directory: {project_path}")
-        sys.exit(1)
-
-    # Validate directory contains .git subdirectory
-    git_dir = project_path / ".git"
-    if not git_dir.exists():
-        logger.error(f"Project directory is not a git repository: {project_path}")
-        sys.exit(1)
-
-    return project_path
 
 
 def prepare_task_tracker(project_dir: Path, llm_method: str) -> bool:
